Logger

+++ src/main/java/com/revature/services/impl/AdminServiceImpl.java
@@ -69,7 +69,10 @@ public class AdminServiceImpl implements AdminService {
 		if(admin.getAdminId() < 1)
 		{
 			logger.warn("Admin ID can't be less than 1");
-		} else if() {
+			return null;
+		} else if(admin.getUserName().equals("")) {
+			logger.warn("User name can't be empty");
+			return null;
 			
 		}
 		return ar.save(admin);
@@ -84,8 +87,9 @@ public class AdminServiceImpl implements AdminService {
 	
 	@Override
 	public Admin updateAdmin(Admin admin) {
-		if(admin.getAdminId() < 0 ) {
-			
+		if(admin.getAdminId() < 1 ) {
+			logger.warn("Admin ID can't be less than 1");
+			return null;
 		}
 		return ar.save(admin);
 	}
@@ -99,6 +103,10 @@ public class AdminServiceImpl implements AdminService {
 	
 	@Override
 	public String deleteAdminById(int id) {
+		if(id < 1) {
+			logger.warn("Admin ID can't be less than 1 ");
+			return null;
+		}
 		ar.deleteById(id);
 		return "Admin with id: " + id + " was deleted.";
 	}
